Add tests.rb, seaparated modules
