+++ enumerables.rb
@@ -96,32 +96,3 @@ module Enumerables
     accumulator
   end
 end
-
-include Enumerables
-
-def multiply_els(array)
-  array.my_inject { |sum, n| sum * n }
-end
-
-test_array = [1, 2, 3, 4, 5, 6, 7, 8]
-test_array.my_each { |x| puts "The number is #{x}" }
-puts
-test_array.my_each_with_index { |val, index| puts "index: #{index} for #{val}" }
-puts
-test_array.my_select { |n| n.even? }
-puts
-puts test_array.my_all { |n| n < 20 }
-puts
-puts test_array.my_none { |n| n < 20 }
-puts
-puts test_array.my_any { |n| n == 20 }
-puts
-puts test_array.my_count { |n| n.even? }
-puts
-square = proc { |x| x**2 }
-# puts test_array.my_map { |x| x**2 }
-puts test_array.my_map(square)
-puts
-puts test_array.my_inject { |sum, n| sum * n }
-puts
-puts multiply_els([2, 4, 5])
